realGA calibration done

+++ tarea3/real_GA/generate_init_pop.m
@@ -4,6 +4,6 @@ function population = generate_init_pop(popSize, lb, ub)
 dim = 10;
 population = {};
 for i = 1:popSize
-    population{i} = lb + (ub - lb) * rand(1, dim);
+    population{i} = double( lb + (ub - lb) * rand(1, dim));
 end
 end
